refactor: improve preferences options widget structure
