+++ lib/widgets/dialog/record_hotkey.dart
