+++ lib/widgets/options/options.dart
@@ -1,0 +1,6 @@
+export 'base_option.dart';
+export 'toggle_option.dart';
+export 'dropdown_option.dart';
+export 'slider_option.dart';
+export 'color_option.dart';
+export 'hotkey_option.dart';
